requirements.txt
Flask==3.0.3
Flask-JWT-Extended==4.6.0
passlib[bcrypt]==1.7.4
pymongo==4.15.2
dnspython==2.8.0
python-dotenv==1.0.1
